phase5-backend-api/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
slowapi
